# swagger.json
{
    "openapi":"3.0.1",
    "info":{
        "version":"1.0.0",
        "title": "API Tallos Manager",
        "description": "API de cadastro de usuários",
        "contact":{
            "name":"David Ric",
            "email":"dric9669@gmail.com"
        }
    
},
"servers":[
    {
    "url":"http://localhost:3000",
    "description": "Local server"
    }
],

"schemes": ["http"],
"paths":{
    "/auth/register":{
        "post":{
            "tags":["Login"],
            "description":"Cadastrar de usuários",
            "operationId":"registroUsuario",
            "security":[{"bearerAuth": [] }],
            "requestBody":{
              "description":"Dados necessários para criar um novo usuário",
              "content":{
                "application/json":{
                    "schema":{
                      "$ref":"#/components/schemas/Cadastra-Usuarios"
                    },
                    "examples":{
                        "Cadatro de Usuarios":{
                            "value":{
                                "name": "David para dev",
                                "email":"david_para_dev_tallos@gmail.com",
                                "password":"12345678",
                                "nivel":1
                            }

                        }
                    }
                   
                }
              }

            },
            "responses":{
                "200":{
                    "description": "Sucess",
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "$ref":"#/components/schemas/Cadastra-Usuarios"
                            
                            }
                        }

                    }
                },
                "400":{
                    "description": "Bad Request",
                    "content":{
                        "application/json":{
                            "schema":{
                                "properties":{
                                    "error":{
                                        "type":"boolean"
                                     },
                                     "message":{
                                        "type":"string"
                                     },
                                     "data":{
                                        "type":"object"
                                     }
                                    }
                            }
                        }

                    }
                },
                "500":{
                    "description": "Server Error",
                    "content":{
                        "application/json":{
                            "schema":{
                            
                            }
                        }

                    }
                }
            }
        }
    },
    "/auth/authenticate":{
        "post":{
            "tags":["Login"],
            "description":"Autenticar usuário",
            "operationId":"autenticarUsuario",
            "security":[{"bearerAuth": [] }],
            "requestBody":{
              "description":"Dados necessários para criar um novo usuário",
              "content":{
                "application/json":{
                    "schema":{
                      "$ref":"#/components/schemas/Autentica-Usuarios"
                    },
                    "examples":{
                        "Cadatro de Usuarios":{
                            "value":{
                                "email":"david_para_dev_tallos@gmail.com",
                                "password":"12345678"
                            }

                        }
                    }
                   
                }
              }

            },
            "responses":{
                "200":{
                    "description": "Sucess",
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "$ref":"#/components/schemas/Autentica-Usuarios"
                            
                            }
                        }

                    }
                },
                "400":{
                    "description": "Bad Request",
                    "content":{
                        "application/json":{
                            "schema":{
                                "properties":{
                                    "error":{
                                        "type":"boolean"
                                     },
                                     "message":{
                                        "type":"string"
                                     },
                                     "data":{
                                        "type":"object"
                                     }
                                    }
                            }
                        }

                    }
                },
                "500":{
                    "description": "Server Error",
                    "content":{
                        "application/json":{
                            "schema":{
                            
                            }
                        }

                    }
                }
            }
        }
       
    },
    "/admin/users":{
    "get":{
      "tags":["Admin/Usuários"],
        "description":"Exibir usuários",
        "operationId":"registroUsuario",
        "security":[{"bearerAuth": [] }],
        "requestBody":{
          "description":"Dados necessários para criar um novo usuário",
          "content":{
            "application/json":{
                "schema":{
                  "$ref":"#/components/schemas/Manipula-Usuarios"
                },
                "examples":{
                    "Cadatro de Usuarios":{
                        "value":{
                            "id": "001",
                            "name": "David para dev",
                            "email":"david_para_dev_tallos@gmail.com",
                            "password":"Pr@45678",
                            "nivel":1
                        }

                    }
                }
               
            }
          }

        },
        "responses":{
            "200":{
                "description": "Sucess",
                "content":{
                    "application/json":{
                        "schema":{
                            "type":"object",
                            "$ref":"#/components/schemas/Manipula-Usuarios"
                        
                        }
                    }

                }
            },
            "400":{
                "description": "Bad Request",
                "content":{
                    "application/json":{
                        "schema":{
                            "properties":{
                                "error":{
                                    "type":"boolean"
                                 },
                                 "message":{
                                    "type":"string"
                                 },
                                 "data":{
                                    "type":"object"
                                 }
                                }
                        }
                    }

                }
            },
            "500":{
                "description": "Server Error",
                "content":{
                    "application/json":{
                        "schema":{
                        
                        }
                    }

                }
            }
           }   }
    },
    "/admin/users/:id":{
        "get":{
         "tags":["Admin/Usuários"],
            "description":"Exibir usuários",
            "operationId":"registroUsuario",
            "security":[{"bearerAuth": [] }],
            "parameters":[
                {
                    "name":"id",
                    "in":"query",
                    "description": "Usuario ID",
                    "required":true
                }
            ],
          
            "responses":{
                "200":{
                    "description": "Sucess",
                    "content":{
                        "application/json":{
                            "schema":{
                                "type":"object",
                                "$ref":"#/components/schemas/Manipula-Usuarios"
                            
                            }
                        }
    
                    }
                },
                "400":{
                    "description": "Bad Request",
                    "content":{
                        "application/json":{
                            "schema":{
                                "properties":{
                                    "error":{
                                        "type":"boolean"
                                     },
                                     "message":{
                                        "type":"string"
                                     },
                                     "data":{
                                        "type":"object"
                                     }
                                    }
                            }
                        }
    
                    }
                },
                "500":{
                    "description": "Server Error",
                    "content":{
                        "application/json":{
                            "schema":{
                            
                            }
                        }
    
                    }
                }
               }   }
        },
        "/admin/users/id":{
            "put":{
             "tags":["Admin/Usuários"],
                "description":"Exibir usuários",
                "operationId":"registroUsuario",
                "security":[{"bearerAuth": [] }],
                "parameters":[
                    {
                        "name":"id",
                        "in":"query",
                        "description": "Usuario ID",
                        "required":true
                    }
                ],  "requestBody":{
                    "description":"Dados necessários para criar um novo usuário",
                    "content":{
                      "application/json":{
                          "schema":{
                            "$ref":"#/components/schemas/Cadastra-Usuarios"
                          },
                          "examples":{
                              "Cadatro de Usuarios":{
                                  "value":{
                                      "name": "David para dev",
                                      "email":"david_para_dev_tallos@gmail.com",
                                      "password":"12345678",
                                      "nivel":1
                                  }
      
                              }
                          }
                         
                      }
                    }
      
                  },
              
                "responses":{
                    "200":{
                        "description": "Sucess",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "type":"object",
                                    "$ref":"#/components/schemas/Manipula-Usuarios"
                                
                                }
                            }
        
                        }
                    },
                    "400":{
                        "description": "Bad Request",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "properties":{
                                        "error":{
                                            "type":"boolean"
                                         },
                                         "message":{
                                            "type":"string"
                                         },
                                         "data":{
                                            "type":"object"
                                         }
                                        }
                                }
                            }
        
                        }
                    },
                    "500":{
                        "description": "Server Error",
                        "content":{
                            "application/json":{
                                "schema":{
                                
                                }
                            }
        
                        }
                    }
                   }   }
            },
        " /admin/users/:id":{
            "delete":{
                 "tags":["Admin/Usuários"],
                "description":"Exibir usuários",
                "operationId":"registroUsuario",
                "security":[{"bearerAuth": [] }],
                "parameters":[
                    {
                        "name":"id",
                        "in":"query",
                        "description": "Usuario ID",
                        "required":true
                    }
                ],
              
                "responses":{
                    "200":{
                        "description": "Sucess",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "type":"object",
                                    "$ref":"#/components/schemas/Manipula-Usuarios"
                                
                                }
                            }
        
                        }
                    },
                    "400":{
                        "description": "Bad Request",
                        "content":{
                            "application/json":{
                                "schema":{
                                    "properties":{
                                        "error":{
                                            "type":"boolean"
                                         },
                                         "message":{
                                            "type":"string"
                                         },
                                         "data":{
                                            "type":"object"
                                         }
                                        }
                                }
                            }
        
                        }
                    },
                    "500":{
                        "description": "Server Error",
                        "content":{
                            "application/json":{
                                "schema":{
                                
                                }
                            }
        
                        }
                    }
                   }   }
            }
},

"components":{
    "schemas":{
        "Cadastra-Usuarios":{
            "type":"object",
            "properties":{
                "name":{
                    "type":"string"
                 },
                 "email":{
                     "type":"string"
                  },
                  "password":{
                     "type":"string"
                  },
                  "nivel":{
                     "type":"number"
                  }
            }
        },
        "Autentica-Usuarios":{
            "type":"object",
            "properties":{

                 "email":{
                     "type":"string"
                  },
                  "password":{
                     "type":"string"
                  }   
            }
        },
        "Manipula-Usuarios":{
            "type":"object",
            "properties":{
                "id":{
                    "type":"string"
                 },
                "name":{
                    "type":"string"
                 },
                 "email":{
                     "type":"string"
                  },
                  "password":{
                     "type":"string"
                  },
                  "nivel":{
                     "type":"number"
                  }
            }
        }
    },
    "securitySchemes":{
        "bearerAuth":{
            "type":"http",
            "scheme":"bearer",
            "bearerFormat":"JWT"
        }
    }

}

}